requirements.txt
pyyaml>=6.0
numpy>=1.24
matplotlib>=3.7
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3
